requirements.txt
flask
flask-cors
ultralytics
opencv-python
numpy
PyJWT
stripe